UPSTREAM: wireguard: allowedips: initialize list head in selftest
The randomized trie tests weren't initializing the dummy peer list head,
resulting in a NULL pointer dereference when used. Fix this by
initializing it in the randomized trie test, just like we do for the
static unit test.

While we're at it, all of the other strings like this have the word
"self-test", so add it to the missing place here.

Fixes: e7096c131e51 ("net: WireGuard secure network tunnel")
Cc: stable@vger.kernel.org
Signed-off-by: Jason A. Donenfeld <Jason@zx2c4.com>
Signed-off-by: David S. Miller <davem@davemloft.net>
(cherry picked from commit 46cfe8eee285cde465b420637507884551f5d7ca)
Signed-off-by: Greg Kroah-Hartman <gregkh@google.com>
Change-Id: I78bee0956e60944a39c3426e0e55675666e86123
Signed-off-by: xiaoleGun <1592501605@qq.com>